+++ clients/src/main/java/org/apache/kafka/common/network/SslTransportLayer.java
@@ -754,7 +754,7 @@ public class SslTransportLayer implements TransportLayer {
     public void removeInterestOps(int ops) {
         if (!key.isValid())
             throw new CancelledKeyException();
-        else if (state != State.READY && state != State.HANDSHAKE_FAILED)
+        else if (state != State.READY)
             throw new IllegalStateException("handshake is not completed");
 
         key.interestOps(key.interestOps() & ~ops);
