+++ clients/src/main/java/org/apache/kafka/common/security/authenticator/SaslClientAuthenticator.java
@@ -322,10 +322,6 @@ public class SaslClientAuthenticator implements Authenticator {
         return saslState == SaslState.COMPLETE;
     }
 
-    @Override
-    public void handleAuthenticationFailure() throws IOException {
-    }
-
     public void close() throws IOException {
         if (saslClient != null)
             saslClient.dispose();
